build: support nms while using MockBukkit in tests
This is tricky because the paper-server jar includes META-INF
service files for real service implementations. Those implementations
conflict with the fakes and mocks defined by MockBukkit. So, to have
both on the classpath without conflicts, a build transform is used to
copy the paper-server jar and delete the relevant entries from its
META-INF/services folder.